Use the game executable's directory as working directory
This is how Windows normally starts applications. This fixes the issue with
games not starting correctly or acting weird, because by default the games
were started with the working directory of the parent process, which is the
Vapor Store in this case.

This creates a bit of duplicate code blocks. If you have an idea to refactor
this, please feel free to do so or give me instructions.